+++ src/entities/Employer/MyStore.tsx
@@ -1,0 +1,14 @@
+import EmpolyerEmptyData from './UI/EmpolyerEmptyData';
+
+const MyStore = () => {
+  return (
+    <>
+      <EmpolyerEmptyData
+        title='내 가게'
+        comment='내 가게를 소개하고 공고도 등록해 보세요'
+        content='가게 등록하기'
+      />
+    </>
+  );
+};
+export default MyStore;
