+++ src/entities/Employer/RegisteredRecruitment.tsx
@@ -1,0 +1,15 @@
+import EmpolyerEmptyData from './UI/EmpolyerEmptyData';
+
+const RegisteredRecruitment = () => {
+  return (
+    <>
+      <EmpolyerEmptyData
+        title='등록한 공고'
+        comment='공고를 등록해 보세요'
+        content='공고 등록하기'
+      />
+    </>
+  );
+};
+
+export default RegisteredRecruitment;
