+++ src/entities/Employer/UI/EmpolyerEmptyData.tsx
@@ -7,20 +7,21 @@
  * 버튼 이름을 작성해주세요
  */
 
+import { ReactNode } from 'react';
 import Button from '@/shared/UI/Button';
 
 interface EmployerEmptyDataComponentProps {
   title: string;
   comment: string;
   content: string;
-  onClick?: React.MouseEventHandler<HTMLButtonElement>;
+  button?: ReactNode;
 }
 
 const EmpolyerEmptyData = ({
   title,
   comment,
   content,
-  onClick,
+  button,
 }: EmployerEmptyDataComponentProps) => {
   return (
     <div className='felx flex-col items-start gap-2 px-[237px] py-[60px]'>
@@ -32,9 +33,8 @@ const EmpolyerEmptyData = ({
           <span className='self-stretch text-center font-normal leading-[26px] text-[#111322]'>
             {comment}
           </span>
-          <Button onClick={onClick} type='active'>
-            {content}
-          </Button>
+          {button}
+          <Button type='active'>{content}</Button>
         </div>
       </div>
     </div>
