+++ src/features/Filter/Filter.tsx
@@ -1,3 +1,3 @@
-import Filter from './ui/Filter';
+import Filter from './UI/FilterUi';
 
 export default Filter;
