--- src/features/Filter/model/model.tsx
@@ -1,1 +1,0 @@
-// api 타입 관련 객체 모음
