--- src/features/Filter/model/store.tsx
@@ -1,1 +1,0 @@
-// api 관련 패칭 함수 모음
