--- src/features/MyStore/MyStore.tsx
@@ -1,4 +1,0 @@
-const MyStore = () => {
-  return <div></div>;
-};
-export default MyStore;
