--- src/features/RegisteredRecruitment/RegisteredRecruitment.tsx
@@ -1,5 +1,0 @@
-const RegisteredRecruitment = () => {
-  return <div></div>;
-};
-
-export default RegisteredRecruitment;
